+++ budget/src/main/java/com/taubel/budget/entities/Budget.java
@@ -32,6 +32,6 @@ public class Budget {
     private Month month;
 
     @Column(name = "year", nullable = false)
-    private int year;
+    private Short year;
 
 }
